# jwt: Cache the bcrypt password hash.