@harshavardhana 2017-01-10
Can you review this @balamurugana ? - i removed the salt usage. There were some bugs on windows.